Fixed clang-tidy warnings [skip ci]

+++ include/bayestest.hpp
@@ -44,7 +44,7 @@ inline double prob_c_beats_ab(int alpha_a, int beta_a, int alpha_b, int beta_b, 
 
   double abc = beta_a + beta_b + beta_c;
   std::vector<double> logbeta_ac_i_j;
-  logbeta_ac_i_j.reserve(static_cast<size_t>(alpha_a + alpha_b));
+  logbeta_ac_i_j.reserve(static_cast<size_t>(alpha_a) + static_cast<size_t>(alpha_b));
 
   for (int i = 0; i < alpha_a + alpha_b; i++) {
     logbeta_ac_i_j.push_back(logbeta(alpha_c + i, abc));
@@ -54,7 +54,7 @@ inline double prob_c_beats_ab(int alpha_a, int beta_a, int alpha_b, int beta_b, 
     double sum_i = -std::log(beta_a + i) - logbeta(1 + i, beta_a) - logbeta_ac_bc;
 
     for (int j = 0; j < alpha_b; j++) {
-      total += std::exp(sum_i + logbeta_ac_i_j.at(static_cast<size_t>(i + j)) - log_bb_j_logbeta_j_bb.at(static_cast<size_t>(j)));
+      total += std::exp(sum_i + logbeta_ac_i_j.at(static_cast<size_t>(i) + static_cast<size_t>(j)) - log_bb_j_logbeta_j_bb.at(static_cast<size_t>(j)));
     }
   }
 
@@ -85,7 +85,7 @@ inline double prob_d_beats_abc(int alpha_a, int beta_a, int alpha_b, int beta_b,
 
   double abcd = beta_a + beta_b + beta_c + beta_d;
   std::vector<double> logbeta_bd_i_j_k;
-  logbeta_bd_i_j_k.reserve(static_cast<size_t>(alpha_a + alpha_b + alpha_c));
+  logbeta_bd_i_j_k.reserve(static_cast<size_t>(alpha_a) + static_cast<size_t>(alpha_b) + static_cast<size_t>(alpha_c));
 
   for (int i = 0; i < alpha_a + alpha_b + alpha_c; i++) {
     logbeta_bd_i_j_k.push_back(logbeta(alpha_d + i, abcd));
@@ -98,7 +98,7 @@ inline double prob_d_beats_abc(int alpha_a, int beta_a, int alpha_b, int beta_b,
       double sum_j = sum_i - log_bb_j_logbeta_j_bb.at(static_cast<size_t>(j));
 
       for (int k = 0; k < alpha_c; k++) {
-        total += std::exp(sum_j + logbeta_bd_i_j_k.at(static_cast<size_t>(i + j + k)) - log_bc_k_logbeta_k_bc.at(static_cast<size_t>(k)));
+        total += std::exp(sum_j + logbeta_bd_i_j_k.at(static_cast<size_t>(i) + static_cast<size_t>(j) + static_cast<size_t>(k)) - log_bc_k_logbeta_k_bc.at(static_cast<size_t>(k)));
       }
     }
   }
